# gamification-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamification API
  description: description
security:
  - apiKeyAuth: []
paths:
...
  /badges/{name}:
    get:
      description: get one badge, by name
      operationId: getBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: badge not found
    put:
      description: update a badge, by name
      operationId: putBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
        - in: query
          name: usable
          schema:
            type: boolean
            default: true
          description: if the badge has to be desactivated pass this value with false
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
        description: pass an empty body if you would like to only modify the usable value
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
        '400':
          description: request body was incorrect
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '409':
          description: badge's name conflict

    delete:
      description: delete the specified badge
      operationId: removeBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Badge Removed
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Badge not found
...
  /rules:
    post:
      description: create a new rule
      operationId: createRule
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
      responses:
        '201':
          description: New rule created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          description: Missing important fields
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: If badge name sent is not known or unusable
        '409':
          description: Rule duplication (same API-KEY, event type and pointscale id)
    get:
      description: get the list of all rules
      operationId: getRules
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleInfo'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
  /rules/{id}:
    get:
      description: get one rules, by id
      operationId: getRule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Rule'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Rule not found
    delete:
      description: delete the specified rule
      operationId: removeRule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: rule removed
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: rule not found
...
  /pointscales:
    post:
      description: create a pointscale
      operationId: createPointScale
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PointScale'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                type: string
        '400':
          description: request body was incorrect
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: usable badge not found
    get:
      description: get the list of all pointscales
      operationId: getPointScales
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointScaleInfo'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
  /pointscales/{id}:
    get:
      description: get one pointscale, by id
      operationId: getPointScale
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PointScale'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: pointscale not found
    delete:
      description: delete the specified pointscale
      operationId: removePointScale
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: pointscale Removed
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: pointscale not found
...
  /rankings/byPoints/pointscales/{id}:
    get:
      description: get the user rankings by points for a pointscale
      operationId: getRankingsByPointScalesPoints
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 0
            default: 10
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PaginatedPointsRankings'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: pointscale not found
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
...
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
...
    Badge:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    Rule:
      type: object
      required:
        - name
        - description
        - eventType
        - pointsToAdd
        - pointScaleId
      properties:
        name:
          type: string
        description:
          type: string
        eventType:
          type: string
        pointsToAdd:
          type: number
          format: double
        badgeName:
          type: string
        pointScaleId:
          type: integer
    RuleInfo:
      type: object
      required:
        - id
        - name
        - description
        - eventType
        - pointsToAdd
        - pointScaleId
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        eventType:
          type: string
        pointsToAdd:
          type: number
          format: double
        badgeName:
          type: string
        pointScaleId:
          type: integer
...
    PointScaleInfo:
      type: object
      required:
        - id
        - stages
      properties:
        id:
          type: integer
        stages:
          type: array
          items:
            $ref: '#/components/schemas/Stage'
    PointScale:
      type: object
      required:
        - stages
      properties:
        stages:
          type: array
          items:
            $ref: '#/components/schemas/Stage'
    Stage:
      type: object
      required:
        - points
        - badge
      properties:
        points:
          type: number
          format: double
        badge:
          $ref: '#/components/schemas/Badge'
    PaginatedPointsRankings:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          type: array
          items:
            $ref: '#/components/schemas/PointsRanking'
...
    PointsRanking:
      type: object
      properties:
        userId:
          type: string
        points:
          type: number
          format: double
...
    Pagination:
      type: object
      required:
        - numberOfItems
        - page
      properties:
        numberOfItems:
          type: integer
          format: int64
        page:
          type: integer
        previous:
          type: string
        next:
          type: string
...